pyproject.toml
[tool.poetry]
name = "telegram bot"
version = "0.1.0"
description = ""
authors = ["Dev Modi <dm.dev.modi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Telethon = "^1.23.0"
pymongo = "^3.12.0"
dnspython = "^2.1.0"
python-dotenv = "^0.18.0"

[tool.poetry.dev-dependencies]
pylint = "^2.9.3"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
